+++ src/ssr/pie.ts
@@ -1,5 +1,8 @@
 import { createChart } from '@antv/g2-ssr';
-import { type PieOptions } from '../type';
+import { type PieProps } from '../export';
+import { type BaseChartConfig } from './type';
+
+export type PieOptions = BaseChartConfig & { type: 'pie' } & PieProps;
 
 export async function Pie(options: PieOptions) {
   const { data, title, width, height, innerRadius } = options;
